[pr][feature] PAM related functionality

+++ ccore/src/cluster/silhouette_ksearch.cpp
@@ -106,7 +106,7 @@ void silhouette_ksearch::process(const dataset & p_data, silhouette_ksearch_data
         silhouette().process(p_data, clusters, result);
 
         const auto & scores = result.get_score();
-        const double score = std::accumulate(scores.begin(), scores.end(), 0.0) / scores.size();
+        const double score = std::accumulate(scores.begin(), scores.end(), 0.0) / static_cast<double>(scores.size());
         p_result.scores().push_back(score);
 
         if (score > p_result.get_score()) {
